Vector ctors from string for CharacterVector only
When it is called for other vector types, a clear error is given.

I have to adjust the included header orders to make it work. If should be safe,
but I'm not hundred percentage sure, due to the complexity of code.

Also move the default template argument to the declaration of vector.